improve form and list supported langs
